requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.1+cpu
transformers==4.44.2
chromadb==1.5.9
langchain==0.2.16
langchain-community==0.2.16
huggingface_hub==0.36.2
fastapi==0.115.5
uvicorn==0.32.1
gradio==5.6.0
pillow==10.4.0
python-dotenv==1.2.2
sqlalchemy==2.0.49
numpy==1.26.4
